Light shader

+++ src/color.rs
@@ -1,12 +1,12 @@
 #[derive(Clone, Copy)]
 pub struct Vec3 {
-    pub e: [f64; 3],
+    pub e: [f32; 3],
 }
 
 pub type Color = Vec3;
 
 impl Vec3 {
-    pub fn new(r: f64, g: f64, b: f64) -> Self {
+    pub fn new(r: f32, g: f32, b: f32) -> Self {
         Self { e: [r, g, b] }
     }
 }
